for_tips/requirements.txt
# Core
torch>=2.2.0
torchvision>=0.17.0
numpy>=1.24.0
opencv-python>=4.9.0
Pillow>=10.0.0

# YOLO + Detection
ultralytics>=8.2.0
supervision>=0.18.0

# Grounding + Segmentation (auto labeling)
groundingdino-py
segment-anything-2

# Scene Graph
torch-geometric>=2.5.0

# GUI Labeling
streamlit>=1.33.0

# Data
h5py>=3.10.0
pandas>=2.0.0
pyyaml>=6.0

# Training utils
tensorboard>=2.16.0
huggingface-hub>=0.22.0
matplotlib>=3.8.0
seaborn>=0.13.0
